added daily summary and its associated files; refactored other related files as needed

+++ app/src/main/java/com/example/nutrimons/database/DateDataDao.java
@@ -41,7 +41,7 @@ public interface DateDataDao {
 
 
     @Query("SELECT water FROM dateData WHERE date LIKE :search")
-    List<Double> findWaterByDate(String search);
+    List<Float> findWaterByDate(String search);
 
 
     // UPDATE QUERIES
